Revert "Fixed alignment in the rank card and improved trunacation"
