+++ bot/extensions/levelling/commands.py
@@ -230,13 +230,14 @@ class Levelling(commands.Cog):
         xp_offset_x -= xp_text_size[2] - xp_text_size[0]
         draw.text((xp_offset_x, xp_offset_y), text, font=self.small_font, fill="#fff")
 
-        if len(username) >= 18:
+        if len(username) >= 15:
             # Truncating the name
             username = username[:15] + "..."
 
         text_bbox = draw.textbbox((0, 0), username, font=self.medium_font)
+        text_offset_x = bar_offset_x - 10
         text_offset_y = bar_offset_y - (text_bbox[3] - text_bbox[1]) - 20
-        draw.text((bar_offset_x, text_offset_y), username, font=self.medium_font, fill="#fff")
+        draw.text((text_offset_x, text_offset_y), username, font=self.medium_font, fill="#fff")
 
         # create copy of background
         background = self.background.copy()
@@ -272,6 +273,8 @@ class Levelling(commands.Cog):
         """
 
         record = await LevellingUser.pool.fetchrow(query, interaction.guild.id, member.id)
+
+        log.info(record)
 
         if record.total_xp is None:
             if member.id == interaction.user.id:
