+++ cli.py
@@ -137,10 +137,10 @@ async def main(ctx):
         "bot.extensions.tags",
         "bot.extensions.levelling",
         "bot.extensions.persistent_roles",
-        "bot.extensions.poll",
         "bot.cogs._help",
         "bot.cogs.clashofcode",
         "bot.cogs.roles",
+        "bot.cogs.poll",
     )
 
     intents = discord.Intents.all()
